+++ src/common/src/telemetry/report.rs
@@ -12,21 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::sync::{Arc, OnceLock};
+use std::sync::Arc;
 
-use prost::Message;
-use risingwave_pb::telemetry::{
-    EventMessage as PbEventMessage, PbTelemetryDatabaseObject,
-    TelemetryEventStage as PbTelemetryEventStage,
+pub use risingwave_telemetry_event::{
+    current_timestamp, post_telemetry_report_pb, TELEMETRY_REPORT_URL, TELEMETRY_TRACKING_ID,
 };
 use tokio::sync::oneshot::Sender;
 use tokio::task::JoinHandle;
 use tokio::time::{interval, Duration};
 use uuid::Uuid;
 
-use super::{current_timestamp, Result, TELEMETRY_REPORT_INTERVAL, TELEMETRY_REPORT_URL};
+use super::{Result, TELEMETRY_REPORT_INTERVAL};
 use crate::telemetry::pb_compatible::TelemetryToProtobuf;
-use crate::telemetry::post_telemetry_report_pb;
 
 #[async_trait::async_trait]
 pub trait TelemetryInfoFetcher {
@@ -46,8 +43,6 @@ pub trait TelemetryReportCreator {
 
     fn report_type(&self) -> &str;
 }
-
-static TELEMETRY_TRACKING_ID: OnceLock<String> = OnceLock::new();
 
 pub async fn start_telemetry_reporting<F, I>(
     info_fetcher: Arc<I>,
@@ -125,100 +120,4 @@ where
         }
     });
     (join_handle, shutdown_tx)
-}
-
-pub fn report_event_common(
-    event_stage: PbTelemetryEventStage,
-    event_name: &str,
-    catalog_id: i64,
-    connector_name: Option<String>,
-    object: Option<PbTelemetryDatabaseObject>,
-    attributes: Option<jsonbb::Value>, // any json string
-    node: String,
-) {
-    let event_tracking_id: String;
-    if let Some(tracking_id) = TELEMETRY_TRACKING_ID.get() {
-        event_tracking_id = tracking_id.to_string();
-    } else {
-        tracing::info!("Telemetry tracking_id is not set, event reporting disabled");
-        return;
-    }
-
-    request_to_telemetry_event(
-        event_tracking_id,
-        event_stage,
-        event_name,
-        catalog_id,
-        connector_name,
-        object,
-        attributes,
-        node,
-        false,
-    );
-}
-
-fn request_to_telemetry_event(
-    tracking_id: String,
-    event_stage: PbTelemetryEventStage,
-    event_name: &str,
-    catalog_id: i64,
-    connector_name: Option<String>,
-    object: Option<PbTelemetryDatabaseObject>,
-    attributes: Option<jsonbb::Value>, // any json string
-    node: String,
-    is_test: bool,
-) {
-    let event = PbEventMessage {
-        tracking_id,
-        event_time_sec: current_timestamp(),
-        event_stage: event_stage as i32,
-        event_name: event_name.to_string(),
-        connector_name,
-        object: object.map(|c| c as i32),
-        catalog_id,
-        attributes: attributes.map(|a| a.to_string()),
-        node,
-        is_test,
-    };
-    let report_bytes = event.encode_to_vec();
-
-    tokio::spawn(async move {
-        const TELEMETRY_EVENT_REPORT_TYPE: &str = "event";
-        let url = (TELEMETRY_REPORT_URL.to_owned() + "/" + TELEMETRY_EVENT_REPORT_TYPE).to_owned();
-        post_telemetry_report_pb(&url, report_bytes)
-            .await
-            .unwrap_or_else(|e| tracing::info!("{}", e))
-    });
-}
-
-#[cfg(test)]
-mod test {
-
-    use super::*;
-
-    #[ignore]
-    #[tokio::test]
-    async fn test_telemetry_report_event() {
-        let event_stage = PbTelemetryEventStage::CreateStreamJob;
-        let event_name = "test_feature";
-        let catalog_id = 1;
-        let connector_name = Some("test_connector".to_string());
-        let object = Some(PbTelemetryDatabaseObject::Source);
-        let attributes = None;
-        let node = "test_node".to_string();
-
-        request_to_telemetry_event(
-            "7d45669c-08c7-4571-ae3d-d3a3e70a2f7e".to_string(),
-            event_stage,
-            event_name,
-            catalog_id,
-            connector_name,
-            object,
-            attributes,
-            node,
-            true,
-        );
-
-        tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-    }
 }
