+++ src/common/telemetry_event/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "telemetry_telemetry_event"
+name = "risingwave_telemetry_event"
 version = { workspace = true }
 edition = { workspace = true }
 homepage = { workspace = true }
@@ -7,4 +7,26 @@ keywords = { workspace = true }
 license = { workspace = true }
 repository = { workspace = true }
 
+[package.metadata.cargo-machete]
+ignored = ["workspace-hack"]
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["workspace-hack"]
+
 [dependencies]
+jsonbb = { workspace = true }
+prost = { workspace = true }
+reqwest = { version = "0.12.2", features = ["json"] }
+risingwave_pb = { workspace = true }
+thiserror = "1"
+thiserror-ext = { workspace = true }
+tokio = { version = "0.2", package = "madsim-tokio", features = [
+    "rt",
+    "rt-multi-thread",
+    "sync",
+    "macros",
+    "time",
+    "signal",
+] }
+tracing = "0.1"
+uuid = { version = "1", features = ["v4"] }
