+++ src/stream/src/telemetry.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::telemetry::report::report_event_common;
+use risingwave_common::telemetry::report_event_common;
 use risingwave_pb::telemetry::{PbTelemetryDatabaseObject, PbTelemetryEventStage};
 
 const TELEMETRY_COMPUTE_REPORT_TYPE: &str = "compute";
